requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmudict==1.0.32
cryptography==44.0.2
dataclasses-json==0.6.7
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.5.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.8
langgraph==0.3.29
langgraph-api==0.1.3
langgraph-checkpoint==2.0.24
langgraph-cli==0.2.3
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.4
langgraph-sdk==0.1.61
langsmith==0.3.30
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==79.0.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.1
structlog==25.2.0
tavily-python==0.5.4
tenacity==9.1.2
textstat==0.7.5
tiktoken==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.5
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
